+++ CadEditor/settings_banana_prince/Settings_Banana Prince-1-1.cs
@@ -103,9 +103,9 @@ public class Data
   {
       int width =  ConfigScript.getLevelWidth(curActiveLayout);
       int height = ConfigScript.getLevelHeight(curActiveLayout);
-      byte[] layer = new byte[width * height];
+      int[] layer = new int[width * height];
       for (int i = 0; i < width * height; i++)
-          layer[i] = (byte)i;
+          layer[i] = i;
       return new LevelLayerData(width, height, layer, null, null);
   }
